[import/no-restricted-paths]: Correctly resolve symlinks in file path
I have been using this rule for a while in a monorepo and turns out it never worked because it wasn't resolving the symlinks inside node_modules correctly. I am not sure if this is the right location and if I should use the sync or async version of this function but this does solve my problem. Would like to hear what you think of this solution.